refactor: move open_path to flox-rust-sdk
Upcoming `[include]` merge operations that will be implemented in
flox-rust-sdk will need to open included environments, so move
`open_path` to flox-rust-sdk. This also requires moving
`UninitializedEnvironment`.
